Only edit configuration's collector if not in use
Only allow users to edit the collector associated to a configuration if
this configuration is not currently used by any sidecars.

The reason for this restriction is to avoid any problems in sidecars
using that configuration, while encouraging users to clone
configurations as a safer way of testing new collectors.